pyproject.toml
[project]
name = "chatter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agno>=1.7.0",
    "dotenv>=0.9.9",
    "elevenlabs>=2.5.0",
    "google-genai>=1.23.0",
    "openai>=1.93.0",
    "python-dotenv>=1.1.1",
    "python-vlc>=3.0.21203",
]
